# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ToNote App OpenApi API Documentation",
        "description": "ToNote App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "schneidershades@gmail.com"
        },
        "version": "3.0"
    },
    "servers": [
        {
            "url": "http://getwordprime.herokuapp.com/",
            "description": "Staging API server"
        }
    ],
    "paths": {
...
        "/api/v1/marketplace-saved": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "userMarketplaceSaved",
                "description": "userMarketplaceSaved",
                "operationId": "userMarketplaceSaved",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Post userMarketplacesaved",
                "description": "Post userMarketplacesaved",
                "operationId": "postuserMarketplacesaved",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMarketplaceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/marketplace-saved/{id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Show an marketplacesaved",
                "description": "Show an marketplacesaved",
                "operationId": "showMarketplacesaved",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "marketplace ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Delete a marketplace saved",
                "description": "Delete a marketplace saved",
                "operationId": "deleteMarketplaceSaved",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "marketplace ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/script-type-categories": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get Script-type",
                "description": "Get script-types",
                "operationId": "allScriptTypeCategory",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/script-type-categories/{id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Show script type",
                "description": "Show script type",
                "operationId": "showScriptType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Script Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreMarketplaceRequest": {
                "title": "Store Marketplace Request Fields",
                "description": "Store Marketplace request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "freelancer_ad_id": {
                        "title": "project freelancer_ad_id",
                        "description": "freelancer_ad_id of the marketplace",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}